Create Crudxml.php
PHP code creates a web page that allows you to list items from an XML file, add new items, and delete existing items. Make sure you have a file called "dados.xml" in the same directory as.This is just a basic example and is not suitable for production use as it does not include adequate data validation or robust security measures.